+++ fiddles/angular2-cli/fiddle-0017-PersistentLinkList/src/app/services/links.service.ts
@@ -53,7 +53,7 @@ export class LinksService {
       (observer: any) => this._errorObserver = observer
     ).share();
 
-    this.request('assets/json/links.json');
+    this.request('links.json');
 
   }