+++ fiddles/angular2-seeder/fiddle-0003-PersistentLinkList/src/client/app/interfaces/link.ts
@@ -1,0 +1,16 @@
+
+export interface Link {
+  id: number;
+  sortId: number;
+  helpLink: string;
+  helpUrl: string;
+  helpIcon: string;
+  helpTitle: string;
+  title: string;
+  icon: string;
+  url: string;
+  text: string;
+  selected?: boolean;
+  hasHelp?: boolean;
+  json?: string;
+}